Clean up the area
(hi hi)

I wanted to figure out what can be done with area.dat.
Apparently something is missing for that; but I can see that
it seems to just map callsign prefixes to smaller areas of countries,
only for a couple of countries. So it seems of limited use until
more code is written (and perhaps more data found as well?)
So ifdef it out for now to reduce code size.