Expect data files in ../share/clu/ relative to executable
That should work ok for most installations, as well as for running
in-place after checking out the git repo and building it.  So we move
the data files accordingly.

Also save the .url file every time we load the web page, so that
the timestamp is updated, so that if this script is run frequently,
we won't check again until the next day.